Implement file error

+++ src/exec/heredoc.c
@@ -6,7 +6,7 @@
 /*   By: cpak <cpak@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/28 04:58:58 by cpak              #+#    #+#             */
-/*   Updated: 2022/05/09 15:13:27 by cpak             ###   ########seoul.kr  */
+/*   Updated: 2022/05/09 16:11:18 by cpak             ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,13 +41,15 @@ static void	write_prompt_to_heredoc(char *delimiter)
 	close(fd);
 }
 
-void	exec_heredoc(char *command[], int out_fd)
+int	exec_heredoc(char *command[], int out_fd)
 {
 	restore_ori_stdin();
 	restore_ori_stdout();
 	write_prompt_to_heredoc(command[0]);
 	dup2(out_fd, STDOUT_FILENO);
 	close(out_fd);
-	print_file(".heredoc");
+	if (!print_file(".heredoc"))
+		return (0);
 	unlink(".heredoc");
+	return (1);
 }
